requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
langchain-community>=0.0.10
openai>=1.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
